keep and vault clicks complete modal show and hide working

+++ Keepr.client/src/components/KeepComponent.vue
@@ -1,7 +1,7 @@
 <template>
   <section @click="setActiveKeep(keep.id)" id="name"
     class="row m-1 main elevation-3 rounded  justify-content-between selectable"
-    :style="{ backgroundImage: `url('${keep.img}')` }" data-bs-toggle="modal" data-bs-target="#KeepModal">
+    :style="{ backgroundImage: `url('${keep.img}')` }">
     <section class="col-12 text-end">
       <button @click.stop="removeKeep(keep.id)" v-if="account.id == keep.creatorId"
         class="btn btn--outline text-danger mdi mdi-delete"></button>
@@ -27,6 +27,7 @@ import Pop from "../utils/Pop.js";
 import { logger } from "../utils/Logger.js";
 import { keepsService } from "../services/KeepsService.js";
 import { router } from "../router.js";
+import { Modal } from "bootstrap";
 export default {
   props: ({
     keep: { type: Object, required: true }
@@ -39,8 +40,8 @@ export default {
       account: computed(() => AppState.account),
       async setActiveKeep(keepId) {
         try {
-          await Pop.confirm("Are you sure you want to delete this Keep?")
           await keepsService.setActiveKeep(keepId);
+          Modal.getOrCreateInstance("#KeepModal").show();
         } catch (error) {
           Pop.error(error)
           logger.log(error)
@@ -49,6 +50,9 @@ export default {
 
       async removeKeep(keepId) {
         try {
+          Modal.getOrCreateInstance("#KeepModal").hide();
+          const confirm = await Pop.confirm("Are you sure you want to delete this Keep?")
+          if (!confirm) { return }
           await keepsService.removeKeep(keepId);
         } catch (error) {
           Pop.error(error);
