+++ Keepr.client/src/components/VaultComponent.vue
@@ -1,15 +1,14 @@
 <template>
-  <router-link :to="{ name: 'Vault', params: { id: vault.id } }">
 
-    <section class="row m-1 main elevation-3 rounded" :style="{ backgroundImage: `url('${vault.img}')` }">
-      <section class="col-12 text-end">
-        <button @click.stop="removeVault(vault.id)" v-if="account.id == vault.creatorId"
-          class="btn btn--outline text-danger mdi mdi-delete"></button>
-      </section>
-      <section id="name" class="col-12 selectable d-flex justify-content-between align-items-end">
+  <section class="row m-1 main elevation-3 rounded selectable" :style="{ backgroundImage: `url('${vault.img}')` }">
+    <section class="col-12 text-end removecol">
+      <button @click.stop="removeVault(vault.id)" v-if="account.id == vault.creatorId"
+        class="btn btn--outline text-danger mdi mdi-delete"></button>
+    </section>
+    <router-link class="d-flex" :to="{ name: 'Vault', params: { id: vault.id } }">
+      <section id="name" class="col-12 detailscol d-flex justify-content-between align-items-end">
         <section>
-
-          <p class="m-0 fw-bold fs-5 col-6 text-light">
+          <p class="m-0 fw-bold fs-5 col-6 text-light p-0">
             {{ vault.name }}
           </p>
         </section>
@@ -18,9 +17,9 @@
         </section>
 
       </section>
-    </section>
+    </router-link>
+  </section>
 
-  </router-link>
 </template>
 
 
@@ -41,6 +40,8 @@ export default {
 
       async removeVault(vaultId) {
         try {
+          const confirm = await Pop.confirm("Are you sure you want to delete this Keep?")
+          if (!confirm) { return }
           await vaultsService.removeVault(vaultId);
         } catch (error) {
           Pop.error(error);
@@ -58,6 +59,15 @@ export default {
   background-position: center;
   background-size: cover;
   height: 200px;
+
+  .removecol {
+    height: 25px;
+  }
+
+  .detailscol {
+    height: 180px;
+    width: 100%;
+  }
 }
 
 #name {
